refactor(validation): move email, url and username validation into package #8531
