+++ src/main/java/edu/harvard/iq/dataverse/DataverseContact.java
@@ -16,7 +16,8 @@ import javax.persistence.Index;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-import org.hibernate.validator.constraints.Email;
+
+import edu.harvard.iq.dataverse.validation.ValidateEmail;
 import org.hibernate.validator.constraints.NotBlank;
 
 /**
