+++ src/main/java/edu/harvard/iq/dataverse/authorization/providers/shib/ShibUtil.java
@@ -4,7 +4,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import edu.harvard.iq.dataverse.EMailValidator;
+import edu.harvard.iq.dataverse.validation.EMailValidator;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
