+++ src/main/java/edu/harvard/iq/dataverse/authorization/users/AuthenticatedUser.java
@@ -3,8 +3,7 @@ package edu.harvard.iq.dataverse.authorization.users;
 import edu.harvard.iq.dataverse.Cart;
 import edu.harvard.iq.dataverse.DatasetLock;
 import edu.harvard.iq.dataverse.UserNotification;
-import edu.harvard.iq.dataverse.ValidateEmail;
-import edu.harvard.iq.dataverse.authorization.AccessRequest;
+import edu.harvard.iq.dataverse.validation.ValidateEmail;
 import edu.harvard.iq.dataverse.authorization.AuthenticatedUserDisplayInfo;
 import edu.harvard.iq.dataverse.authorization.AuthenticatedUserLookup;
 import edu.harvard.iq.dataverse.authorization.providers.oauth2.OAuth2TokenData;
