+++ src/main/java/edu/harvard/iq/dataverse/validation/UserNameValidator.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package edu.harvard.iq.dataverse;
+package edu.harvard.iq.dataverse.validation;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
