+++ src/main/java/edu/harvard/iq/dataverse/validation/ValidateEmail.java
@@ -3,31 +3,29 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package edu.harvard.iq.dataverse;
+package edu.harvard.iq.dataverse.validation;
+import static java.lang.annotation.ElementType.*;
+import static java.lang.annotation.RetentionPolicy.*;
 
 import java.lang.annotation.Documented;
-import static java.lang.annotation.ElementType.FIELD;
 import java.lang.annotation.Retention;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
 import java.lang.annotation.Target;
+
 import javax.validation.Constraint;
 import javax.validation.Payload;
-
 /**
  *
- * @author sarahferry
+ * @author skraffmi
  */
-
 @Target({FIELD})
 @Retention(RUNTIME)
-@Constraint(validatedBy = {UserNameValidator.class})
+@Constraint(validatedBy = {EMailValidator.class})
 @Documented
-public @interface  ValidateUserName {
-  String message() default "Failed Validation Username";
+public @interface  ValidateEmail {
+  String message() default "Failed Validation Email Address";
 
   Class<?>[] groups() default {};
 
   Class<? extends Payload>[] payload() default {};
     
 }
-
