+++ src/main/java/edu/harvard/iq/dataverse/validation/ValidateURL.java
@@ -1,5 +1,5 @@
 
-package edu.harvard.iq.dataverse;
+package edu.harvard.iq.dataverse.validation;
 
 import java.lang.annotation.Documented;
 import static java.lang.annotation.ElementType.FIELD;
